market/requirements.txt
requests
beautifulsoup4
lxml
pandas
python-dateutil
yfinance
pytz
